NAD logic is in network-area-diagram-content.js. Zoom and size is currently disabled.
