+++ src/components/diagrams/diagram.js
@@ -74,476 +74,26 @@ import NetworkAreaDiagramContent from './networkAreaDiagram/network-area-diagram
 let initialWidth, initialHeight;
 
 const Diagram = forwardRef((props, ref) => {
-    const [svg, setSvg] = useState(NoSvg);
 
     const dispatch = useDispatch();
-    const { snackError } = useSnackMessage();
-    //const intlRef = useIntlRef();
 
     const svgRef = useRef();
-    const diagramViewerRef = useRef();
 
-    const theme = useTheme();
     const classes = useDiagramStyles();
     const intl = useIntl();
 
     const {
-        openDiagramView,
         minimizeDiagramView,
         togglePinDiagramView,
         closeDiagramView,
     } = useDiagram();
 
-    const network = useSelector((state) => state.network);
-
-    const currentNode = useSelector((state) => state.currentTreeNode);
-
     const fullScreenDiagram = useSelector((state) => state.fullScreenDiagram);
 
-    const notificationIdList = useSelector((state) => state.notificationIdList);
-
-    const [forceState, updateState] = useState(false);
 
     const networkAreaDiagramDepth = useSelector(
         (state) => state.networkAreaDiagramDepth
     );
-
-    const [loadingState, updateLoadingState] = useState(false);
-
-    const [locallySwitchedBreaker, setLocallySwitchedBreaker] = useState();
-
-    const isAnyNodeBuilding = useIsAnyNodeBuilding();
-
-    const [modificationInProgress, setModificationInProgress] = useState(false);
-
-    const errorWidth = MAX_WIDTH_VOLTAGE_LEVEL;
-
-    const [errorMessage, setErrorMessage] = useState('');
-
-    /*const forceUpdate = useCallback(() => {
-        updateState((s) => !s);
-    }, []);*/
-
-    /*useImperativeHandle(
-        ref,
-        () => ({
-            reloadSvg: forceUpdate,
-        }),
-        // Note: forceUpdate doesn't change
-        [forceUpdate]
-    );*/
-
-    // const hasDiagramSizeRemainedTheSame = (
-    //     oldWidth,
-    //     oldHeight,
-    //     newWidth,
-    //     newHeight
-    // ) => {
-    //     return oldWidth === newWidth && oldHeight === newHeight;
-    // };
-
-    // using many useState() calls with literal values only to
-    // easily avoid recomputing stuff when updating with the same values
-    // const [svgPreferredWidth, setSvgPreferredWidth] = useState();
-    // const [svgPreferredHeight, setSvgPreferredHeight] = useState();
-    // const [headerPreferredHeight, setHeaderPreferredHeight] = useState();
-    // const [finalPaperWidth, setFinalPaperWidth] = useState();
-    // const [finalPaperHeight, setFinalPaperHeight] = useState();
-    // const [svgFinalWidth, setSvgFinalWidth] = useState();
-    // const [svgFinalHeight, setSvgFinalHeight] = useState();
-
-    // // Here, the goal is to update the parent's list of heights with the newly computed height of this diagram.
-    // const setDisplayedDiagramHeights = props.setDisplayedDiagramHeights;
-    // useEffect(() => {
-    //     if (finalPaperHeight) {
-    //         setDisplayedDiagramHeights((displayedDiagramHeights) => {
-    //             return [
-    //                 ...displayedDiagramHeights.filter(
-    //                     // We remove any old diagram that matches the current diagram...
-    //                     (diagram) =>
-    //                         diagram.id !== props.diagramId ||
-    //                         diagram.svgType !== props.svgType
-    //                 ),
-    //                 {
-    //                     // ...and then insert the current diagram's height
-    //                     id: props.diagramId,
-    //                     svgType: props.svgType,
-    //                     initialHeight: finalPaperHeight,
-    //                 },
-    //             ];
-    //         });
-    //     }
-    // }, [
-    //     finalPaperHeight,
-    //     setDisplayedDiagramHeights,
-    //     props.diagramId,
-    //     props.svgType,
-    // ]);
-
-    // After getting the SVG, we will calculate the diagram's ideal size
-    // useLayoutEffect(() => {
-    //     const sizes = computePaperAndSvgSizesIfReady(
-    //         fullScreenDiagram?.id,
-    //         props.svgType,
-    //         props.totalWidth,
-    //         props.totalHeight,
-    //         svgPreferredWidth,
-    //         svgPreferredHeight,
-    //         headerPreferredHeight
-    //     );
-    //
-    //     if (sizes) {
-    //         if (
-    //             !fullScreenDiagram?.id &&
-    //             sizes.svgWidth * props.numberToDisplay > props.totalWidth
-    //         ) {
-    //             setSvgFinalWidth(props.totalWidth / props.numberToDisplay);
-    //             setFinalPaperWidth(props.totalWidth / props.numberToDisplay);
-    //
-    //             const adjustedHeight =
-    //                 sizes.svgHeight *
-    //                 (props.totalWidth / props.numberToDisplay / sizes.svgWidth);
-    //
-    //             setSvgFinalHeight(adjustedHeight);
-    //             setFinalPaperHeight(
-    //                 adjustedHeight + (sizes.paperHeight - sizes.svgHeight)
-    //             );
-    //         } else {
-    //             setSvgFinalWidth(sizes.svgWidth);
-    //             setFinalPaperWidth(sizes.paperWidth);
-    //             setSvgFinalHeight(sizes.svgHeight);
-    //             setFinalPaperHeight(sizes.paperHeight);
-    //         }
-    //     }
-    // }, [
-    //     fullScreenDiagram,
-    //     props.totalWidth,
-    //     props.totalHeight,
-    //     props.svgType,
-    //     svgPreferredWidth,
-    //     svgPreferredHeight,
-    //     headerPreferredHeight,
-    //     props.numberToDisplay,
-    //     props.diagramId,
-    // ]);
-
-    const isNodeinNotifs = isNodeInNotificationList(
-        currentNode,
-        notificationIdList
-    );
-
-    useEffect(() => {
-        if (props.svgUrl) {
-            if (!isNodeinNotifs) {
-                const isDiagramTypeSld =
-                    props.svgType === SvgType.VOLTAGE_LEVEL ||
-                    props.svgType === SvgType.SUBSTATION;
-
-                updateLoadingState(true);
-                fetchSvg(props.svgUrl)
-                    .then((data) => {
-                        if (data !== null) {
-                            setSvg({
-                                svg: data.svg,
-                                metadata: isDiagramTypeSld
-                                    ? data.metadata
-                                    : null,
-                                error: null,
-                                svgUrl: props.svgUrl,
-                            });
-                        } else {
-                            setSvg(NoSvg);
-                        }
-                    })
-                    .catch((error) => {
-                        console.error(error.message);
-                        setSvg({
-                            svg: null,
-                            metadata: null,
-                            error: error.message,
-                            svgUrl: props.svgUrl,
-                        });
-                        let msg;
-                        if (error.status === 404) {
-                            msg = `Voltage level ${props.diagramId} not found`; // TODO change this error message
-                        } else {
-                            msg = error.message;
-                        }
-                        snackError({
-                            messageTxt: msg,
-                        });
-                    })
-                    .finally(() => {
-                        updateLoadingState(false);
-                        if (isDiagramTypeSld) {
-                            setLocallySwitchedBreaker();
-                        }
-                    });
-            }
-        } else {
-            setSvg(NoSvg);
-        }
-    }, [
-        props.svgUrl,
-        forceState,
-        snackError,
-        //intlRef,
-        props.diagramId,
-        props.svgType,
-        isNodeinNotifs,
-    ]);
-
-    // shouldResetPreferredSizes doesn't need to be a ref, but it makes the static checks happy
-    // const shouldResetPreferredSizes = useRef();
-    // shouldResetPreferredSizes.current = false;
-    // useLayoutEffect(() => {
-    //     shouldResetPreferredSizes.current = true;
-    //     // Note: these deps must be kept in sync with the ones of the useLayoutEffect where setSvgPreferredWidth and setSvgPreferredHeight
-    //     // are called. Because we want to reset them in all cases, except when only svgFinalWidth and svgFinalHeight have changed
-    //     // so we use the same deps but without svgFinalWidth and svgFinalHeight
-    //     // TODO is there a better way to do this??
-    // }, [
-    //     network,
-    //     svg,
-    //     currentNode,
-    //     props.isComputationRunning,
-    //     isAnyNodeBuilding,
-    //     // equipmentMenu,
-    //     // showEquipmentMenu,
-    //     props.svgType,
-    //     theme,
-    //     props.diagramId,
-    //     ref,
-    //     props.disabled,
-    // ]);
-
-    useLayoutEffect(() => {
-        if (props.disabled) return;
-
-        if (svg.svg) {
-            if (props.svgType === SvgType.NETWORK_AREA_DIAGRAM) {
-                const diagramViewer = new NetworkAreaDiagramViewer(
-                    svgRef.current,
-                    svg.svg,
-                    LOADING_WIDTH, // svgFinalWidth,
-                    LOADING_HEIGHT, // svgFinalHeight,
-                    MAX_WIDTH_NETWORK_AREA_DIAGRAM,
-                    MAX_HEIGHT_NETWORK_AREA_DIAGRAM
-                );
-                // setSvgPreferredHeight(diagramViewer.getHeight());
-                // setSvgPreferredWidth(diagramViewer.getWidth());
-
-                // if original nad size has not changed (nad structure has remained the same), we keep the same zoom
-                if (
-                    diagramViewerRef.current // &&
-                    // hasDiagramSizeRemainedTheSame(
-                    //     diagramViewerRef.current.getOriginalWidth(),
-                    //     diagramViewerRef.current.getOriginalHeight(),
-                    //     diagramViewer.getOriginalWidth(),
-                    //     diagramViewer.getOriginalHeight()
-                    // )
-                ) {
-                    diagramViewer.setViewBox(
-                        diagramViewerRef.current.getViewBox()
-                    );
-                }
-
-                diagramViewerRef.current = diagramViewer;
-            } else {
-                // // props.svgType is of Single Line Diagram type
-                // let viewboxMaxWidth =
-                //     props.svgType === SvgType.VOLTAGE_LEVEL
-                //         ? MAX_WIDTH_VOLTAGE_LEVEL
-                //         : MAX_WIDTH_SUBSTATION;
-                // let viewboxMaxHeight =
-                //     props.svgType === SvgType.VOLTAGE_LEVEL
-                //         ? MAX_HEIGHT_VOLTAGE_LEVEL
-                //         : MAX_HEIGHT_SUBSTATION;
-                // let onNextVoltageCallback =
-                //     !props.isComputationRunning &&
-                //     !isAnyNodeBuilding &&
-                //     !modificationInProgress &&
-                //     !loadingState
-                //         ? handleNextVoltageLevelClick
-                //         : null;
-                // let onBreakerCallback =
-                //     !props.isComputationRunning &&
-                //     !isAnyNodeBuilding &&
-                //     !isNodeReadOnly(currentNode) &&
-                //     !modificationInProgress &&
-                //     !loadingState
-                //         ? (breakerId, newSwitchState, switchElement) => {
-                //               if (!modificationInProgress) {
-                //                   setModificationInProgress(true);
-                //                   updateLoadingState(true);
-                //                   setLocallySwitchedBreaker(switchElement);
-                //                   handleBreakerClick(
-                //                       breakerId,
-                //                       newSwitchState,
-                //                       switchElement
-                //                   );
-                //               }
-                //           }
-                //         : null;
-                // let onEquipmentMenuCallback =
-                //     !props.isComputationRunning &&
-                //     !isAnyNodeBuilding &&
-                //     !modificationInProgress &&
-                //     !loadingState
-                //         ? showEquipmentMenu
-                //         : null;
-                //
-                // let selectionBackColor = theme.palette.background.paper;
-                //
-                // const diagramViewer = new SingleLineDiagramViewer(
-                //     svgRef.current, //container
-                //     svg.svg, //svgContent
-                //     svg.metadata, //svg metadata
-                //     props.svgType,
-                //     svgFinalWidth,
-                //     svgFinalHeight,
-                //     viewboxMaxWidth,
-                //     viewboxMaxHeight,
-                //     onNextVoltageCallback, //callback on the next voltage arrows
-                //     onBreakerCallback, // callback on the breakers
-                //     onEquipmentMenuCallback, //callback on the feeders
-                //     selectionBackColor //arrows color
-                // );
-                //
-                // if (shouldResetPreferredSizes.current) {
-                //     setSvgPreferredHeight(diagramViewer.getHeight());
-                //     setSvgPreferredWidth(diagramViewer.getWidth());
-                // }
-                //
-                // //Rotate clicked switch while waiting for updated sld data
-                // if (locallySwitchedBreaker) {
-                //     const breakerToSwitchDom = document.getElementById(
-                //         locallySwitchedBreaker.id
-                //     );
-                //     if (
-                //         breakerToSwitchDom.classList.value.includes(
-                //             'sld-closed'
-                //         )
-                //     ) {
-                //         breakerToSwitchDom.classList.replace(
-                //             'sld-closed',
-                //             'sld-open'
-                //         );
-                //     } else if (
-                //         breakerToSwitchDom.classList.value.includes('sld-open')
-                //     ) {
-                //         breakerToSwitchDom.classList.replace(
-                //             'sld-open',
-                //             'sld-closed'
-                //         );
-                //     }
-                // }
-                //
-                // //if original sld size has not changed (sld structure has remained the same), we keep the same zoom
-                // if (
-                //     diagramViewerRef.current &&
-                //     hasDiagramSizeRemainedTheSame(
-                //         diagramViewerRef.current.getOriginalWidth(),
-                //         diagramViewerRef.current.getOriginalHeight(),
-                //         diagramViewer.getOriginalWidth(),
-                //         diagramViewer.getOriginalHeight()
-                //     )
-                // ) {
-                //     diagramViewer.setViewBox(
-                //         diagramViewerRef.current.getViewBox()
-                //     );
-                // }
-                //
-                // // on sld resizing, we need to refresh zoom to avoid exceeding max or min zoom
-                // // this is due to a svg.panzoom.js package's behaviour
-                // diagramViewer.refreshZoom();
-                //
-                // diagramViewerRef.current = diagramViewer;
-            }
-        }
-    }, [
-        network,
-        props.diagramId,
-        props.svgUrl,
-        svg,
-        currentNode,
-        props.isComputationRunning,
-        isAnyNodeBuilding,
-        // equipmentMenu,
-        // showEquipmentMenu,
-        props.svgType,
-        theme,
-        ref,
-        // svgFinalHeight,
-        // svgFinalWidth,
-        props.disabled,
-        modificationInProgress,
-        loadingState,
-        locallySwitchedBreaker,
-        networkAreaDiagramDepth,
-        // handleBreakerClick,
-        // handleNextVoltageLevelClick,
-    ]);
-
-    // useLayoutEffect(() => {
-    //     // if (svgFinalWidth != null && svgFinalHeight != null) {
-    //     //     const divElt = svgRef.current;
-    //     //     if (divElt != null) {
-    //     //         const svgEl = divElt.getElementsByTagName('svg')[0];
-    //     //         if (svgEl != null) {
-    //     //             svgEl.setAttribute('width', svgFinalWidth);
-    //     //             svgEl.setAttribute(
-    //     //                 'height',
-    //     //                 props.computedHeight ?? svgFinalHeight
-    //     //             );
-    //     //         }
-    //     //     }
-    //     //     setModificationInProgress(false);
-    //     // }
-    // }, [
-    //     svgFinalWidth,
-    //     svgFinalHeight,
-    //     //TODO, these are from the previous useLayoutEffect
-    //     //how to refactor to avoid repeating them here ?
-    //     svg,
-    //     props.isComputationRunning,
-    //     props.svgType,
-    //     theme,
-    //     // equipmentMenu,
-    //     // showEquipmentMenu,
-    //     locallySwitchedBreaker,
-    //     loadingState,
-    //     modificationInProgress,
-    //     isAnyNodeBuilding,
-    //     network,
-    //     ref,
-    //     fullScreenDiagram,
-    //     props.computedHeight,
-    // ]);
-
-    // let sizeWidth,
-    //     sizeHeight = initialHeight;
-    // if (svg.error) {
-    //     sizeWidth = errorWidth;
-    // } else if (finalPaperWidth != null && finalPaperHeight != null) {
-    //     sizeWidth = finalPaperWidth;
-    //     sizeHeight = finalPaperHeight;
-    // } else if (initialWidth !== undefined || loadingState) {
-    //     sizeWidth = initialWidth;
-    // } else {
-    //     sizeWidth = props.totalWidth; // happens during initialization if initial width value is undefined
-    // }
-    //
-    // if (sizeWidth !== undefined) {
-    //     initialWidth = sizeWidth; // setting initial width for the next SLD.
-    // }
-    // if (sizeHeight !== undefined) {
-    //     initialHeight = sizeHeight; // setting initial height for the next SLD.
-    // }
-    //
-    // if (!fullScreenDiagram?.id && props.computedHeight) {
-    //     sizeHeight = props.computedHeight;
-    // }
 
     /**
      * DIAGRAM CONTROL HANDLERS
@@ -586,7 +136,7 @@ const Diagram = forwardRef((props, ref) => {
      * RENDER
      */
 
-    return !svg.error ? (
+    return (
         <DiagramResizableBox
             align={props.align}
             // height={sizeHeight}
@@ -634,9 +184,6 @@ const Diagram = forwardRef((props, ref) => {
                     </Box>
                 ) : (
                     <Box height={'100%'}>
-                        {errorMessage && (
-                            <Alert severity="error">{errorMessage}</Alert>
-                        )}
 
                         {(props.svgType === SvgType.VOLTAGE_LEVEL ||
                             props.svgType === SvgType.SUBSTATION) && (
@@ -663,12 +210,17 @@ const Diagram = forwardRef((props, ref) => {
                         )}
                         {props.svgType === SvgType.NETWORK_AREA_DIAGRAM && (
                             <NetworkAreaDiagramContent
-                                ref={svgRef}
+                                ref={ref}
                                 loadFlowStatus={props.loadFlowStatus}
+                                diagramId={props.diagramId}
+                                disabled={props.disabled}
+                                fullScreenActive={fullScreenDiagram?.id}
+                                svgUrl={props.svgUrl}
+                                studyUuid={props.studyUuid}
+
                             />
-                            // <div>vide</div>
                         )}
-                        {!loadingState && (
+
                             <DiagramFooter
                                 showCounterControls={
                                     props.svgType ===
@@ -685,13 +237,11 @@ const Diagram = forwardRef((props, ref) => {
                                 onStartFullScreen={onShowFullScreenHandler}
                                 onStopFullScreen={onHideFullScreenHandler}
                             />
-                        )}
+
                     </Box>
                 )}
             </Paper>
         </DiagramResizableBox>
-    ) : (
-        <></>
     );
 });
